requirements-dev.txt
# Dev / CI dependencies

cryptography
snowflake-connector-python
google-cloud-storage
loguru
python-dotenv
pytest-mock
pytest
ruff
